chore: menu animation moved
